Reformatted all files

+++ Telegram.Server/Service/ServiceBase.WriteTo.cs
@@ -1,18 +1,13 @@
 ﻿using System.Text.Json;
 
-namespace Telegram.Server.Service
-{
-    public partial class ServiceBase
-    {
+namespace Telegram.Server.Service;
 
-        public void WriteTo(object data, string filePath)
-        {
-            if (!File.Exists(filePath))
-            {
-                File.Create(filePath);
-            }
-            var json = JsonSerializer.Serialize(data);
-            File.AppendAllText(filePath, json);
-        }
+public partial class ServiceBase
+{
+    public void WriteTo(object data, string filePath)
+    {
+        if (!File.Exists(filePath)) File.Create(filePath);
+        var json = JsonSerializer.Serialize(data);
+        File.AppendAllText(filePath, json);
     }
 }
